requirements.txt
psutil==4.3.1
pyserial==3.2.1
